backend/requirements.txt
flask
torch
pillow
numpy
torchvision
flask_cors